% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{hots_player}
\alias{hots_player}
\title{fetch player information}
\usage{
hots_player(player_id = NULL, region = NULL, battletag = NULL)
}
\arguments{
\item{player_id}{a integer, the hotslogs player id}

\item{region}{a character among 'EU', 'US', 'CN', 'KR'}

\item{battletag}{a character, the Blizzard battletag of a player in "Name#1234" format}
}
\description{
fetch player information
}

